refactor(models): move react encoding op types out of validator
