+++ graphistry/models/surfaces/__init__.py
@@ -21,6 +21,12 @@ from .graphistry_frontend import (  # noqa: F401
     AxisRows,
     RingCategoricalAxis,
     RingContinuousAxis,
+    AxisEncodingKey,
+    AxisEncodingOp,
+    ColorEncodingOp,
+    EncodingOperationKind,
+    IconEncodingOp,
+    ReactEncodingOp,
     APPLY_ENCODINGS_REACT_KEYS,
     APPLY_ENCODINGS_REACT_KEY_SET,
     ApplyEncodingsReactSettingsDict,
@@ -36,6 +42,7 @@ from .graphistry_frontend import (  # noqa: F401
     ReactIconEncodingKey,
     ReactIconEncodingPayload,
     ReactSettingsDict,
+    SizeEncodingOp,
     ReactSizeEncodingKey,
     ReactSizeEncodingPayload,
     graphistry_frontend_contract_version_info,
